Service discovery with master and worker impl have done
